fix formattedNum instead
